Moved FlowsResult to services under same namespace as FlowsFilter
