+++ app/serializers/api/v3/flows/result_serializer.rb
@@ -1,7 +1,7 @@
 module Api
   module V3
     module Flows
-      class FlowsResultSerializer < ActiveModel::Serializer
+      class ResultSerializer < ActiveModel::Serializer
         attributes :data, :include
       end
     end
